pyproject.toml
[project]
name = "docx-editor-skill"
version = "1.0.0"
description = "CLI tool for fine-grained Microsoft Word document manipulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
docx = "docx_editor_skill.cli:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docx_editor_skill"]

[tool.pytest.ini_options]
testpaths = ["tests"]
